+++ internal/server/handshake.go
@@ -1,55 +1,31 @@
 package server
 
 import (
-	"io"
 	"net"
 
-	cbound "havry.dev/havry/hopper/internal/protocol/packets/clientbound"
-	"havry.dev/havry/hopper/internal/protocol/types"
+	sbound "github.com/gavrylenkoIvan/hopper/public/serverbound"
 )
 
 const (
-	ListPacketID int = 0x00
-	PingPacketID int = 0x01
+	StatusState int = 0x01
+	LoginState  int = 0x2
 )
 
-func (h *Hopper) status(conn net.Conn) error {
-	for i := 0; i < 2; i++ {
-		_, packetID, err := ReadPacketInfo(conn)
-		if err != nil {
-			return err
-		}
+func (h *Hopper) handshake(conn net.Conn) error {
+	defer conn.Close()
 
-		switch int(packetID) {
-		case PingPacketID:
-			payload := make([]byte, types.LongBytes)
-			_, err = io.ReadFull(conn, payload)
-			if err != nil {
-				break
-			}
+	// new conn always starts with handshake packet
+	p := new(sbound.Handshake)
+	_, _, err := ReadPacket(conn, p)
+	if err != nil {
+		return err
+	}
 
-			var body []byte
-			body, err = PrependID(PingPacketID, payload)
-			if err != nil {
-				break
-			}
-
-			_, err = WriteRaw(conn, body)
-		case ListPacketID:
-			players := cbound.Players{
-				Max:    h.Config.Motd.MaxPlayers,
-				Online: 0,
-			}
-
-			_, err = WritePacket(conn,
-				ListPacketID,
-				cbound.NewList(h.Config.Motd.Description, players, h.favicon),
-			)
-		}
-
-		if err != nil {
-			return err
-		}
+	switch int(p.NextState) {
+	case StatusState:
+		return h.status(conn)
+	case LoginState:
+		return h.login(conn)
 	}
 
 	return nil
