+++ public/clientbound/packet.go
@@ -1,0 +1,12 @@
+package cbound
+
+import (
+	"io"
+
+	"github.com/gavrylenkoIvan/hopper/public/interfaces"
+)
+
+type Packet interface {
+	interfaces.Packet
+	io.WriterTo
+}
