+++ public/helpers/pubkey.go
@@ -1,4 +1,4 @@
-package protocol
+package helpers
 
 import (
 	"crypto/rand"
