+++ public/interfaces/packet.go
@@ -1,0 +1,5 @@
+package interfaces
+
+type Packet interface {
+	ID() int
+}
