+++ public/serverbound/packet.go
@@ -1,0 +1,7 @@
+package sbound
+
+import "github.com/gavrylenkoIvan/hopper/public/interfaces"
+
+type Packet interface {
+	interfaces.Packet
+}
