+++ public/types/bytearr.go
